rename home to tera-home
