+++ packages/client/hmi-client/src/page/tera-home.vue
